Fix maxPartySize in the config being ignored.
This fixes maxPartySize in the config being ignored, since right now if
you generate an island without any of the extra perms it will default to
4 max members instead of what is set in the config.

This also removes the party.yml and the player.yml which are not used.

It also removes the redundant max party size check in the home command
which breaks the vault permissions for having more party members.