Corrected POM plugin tags per Maven spec:
Project was not correctly building with Maven due to confusion between
<build><plugin> and <build><pluginmanagement>.
The <pluginManagement> tag is used to define plugin properties.
Additional modules still need to reference the plugin for it to be used
by including it in the <build><plugin> area.
See the following links for reference: http://maven.apache.org/pom.html#Plugin_Management
and: http://maven.40175.n5.nabble.com/use-of-plugin-and-pluginmanagement-td3341163.html
and: https://stackoverflow.com/questions/10483180/maven-what-is-pluginmanagement