Update region limits to reflect the new 1.18 world height
Some of the players on our server are complaining that they cannot build below Y0 after the new 1.18 update. I briefly looked into the code and found that the minimum and maximum region height values are hard coded to the old world limits. I propose to change these values to reflect the new world height limits.