Update region limits to reflect the new 1.18 world height (#36)
* Update region limits to reflect the new 1.18 world height

Some of the players on our server are complaining that they cannot build below Y0 after the new 1.18 update. I briefly looked into the code and found that the minimum and maximum region height values are hard coded to the old world limits. I propose to change these values to reflect the new world height limits.

* Get WG region height limits from API (removed hardcoded value)

* Fixed biome changing Y<0 in 1.18 world

* Fixed score calculation in 1.18 worlds

* Reworked world height constants to make the code compaible with 1.18 worlds

* Fixed off by one maximum height value (returned number is exclusive - 320 souhld be 319)

* Fix for FAWE depencency (requires FAWE v2+)