Update to 1.20.4 (and also fix some stuff) (#46)
* Also update the biome IDs in the biome GUI

* Bump to MC 1.20.4

* Fix display.Name item tag in configmenu.yml interfering with I18nUtil.tr in recent version of Paper

It is needed to set the display name of an item in Json format to avoid implicit interpretation of the value of the tag. Here, the 2 items was named "'{'" and "'}'", but the resent version of Paper interpreted it as "{" and "}" without the single quotes, messing with the formatter of I18nUtil.tr() method.